switching back to old data type
git-svn-id: https://ssl.drqueue.org/svn/branches/0.64.x@2817 94d6b6b3-0ee6-0310-ab95-9b4161196ee2

+++ libdrqueue/database.h
@@ -40,8 +40,8 @@
 
 struct load_balancing {
   struct tpol pol[MAXJOBS]; // Priority ordered list of jobs
-  uint64_t next_i;          // Variables used for load balancing
-  uint64_t last_priority;   //
+  uint32_t next_i;          // Variables used for load balancing
+  uint32_t last_priority;   //
 };
 
 struct database {
